Moved initialize_cameras_landmarks_keyframe to sfm
This algorithm has been renamed to just sfm::initialize_camera_landmarks.
The old core algorithm with this name is deprecated and will be removed.
