+++ arrows/core/CMakeLists.txt
@@ -31,7 +31,6 @@ set( plugin_core_headers
   handle_descriptor_request_core.h
   hierarchical_bundle_adjust.h
   initialize_cameras_landmarks.h
-  initialize_cameras_landmarks_keyframe.h
   initialize_object_tracks_threshold.h
   keyframe_selector_basic.h
   match_features_fundamental_matrix.h
@@ -65,7 +64,6 @@ set( plugin_core_headers
   match_matrix.h
   metrics.h
   projected_track_set.h
-  sfm_utils.h
   track_features_core.h
   track_set_impl.h
   transform.h
@@ -114,7 +112,6 @@ set( plugin_core_sources
   handle_descriptor_request_core.cxx
   hierarchical_bundle_adjust.cxx
   initialize_cameras_landmarks.cxx
-  initialize_cameras_landmarks_keyframe.cxx
   initialize_object_tracks_threshold.cxx
   keyframe_selector_basic.cxx
   match_features_fundamental_matrix.cxx
@@ -148,7 +145,6 @@ set( plugin_core_sources
   match_matrix.cxx
   metrics.cxx
   projected_track_set.cxx
-  sfm_utils.cxx
   track_features_core.cxx
   track_set_impl.cxx
   transform.cxx
