+++ arrows/sfm/initialize_cameras_landmarks.h
@@ -30,34 +30,34 @@
 
 /**
  * \file
- * \brief Header for core camera and landmark initialization algorithm
+ * \brief Header for SFM camera and landmark initialization algorithm
  */
 
-#ifndef KWIVER_ARROWS_CORE_INITIALIZE_CAMERAS_LANDMARKS_KEYFRAME_H_
-#define KWIVER_ARROWS_CORE_INITIALIZE_CAMERAS_LANDMARKS_KEYFRAME_H_
+#ifndef KWIVER_ARROWS_SFM_INITIALIZE_CAMERAS_LANDMARKS_H_
+#define KWIVER_ARROWS_SFM_INITIALIZE_CAMERAS_LANDMARKS_H_
 
-#include <arrows/core/kwiver_algo_core_export.h>
+#include <arrows/sfm/kwiver_algo_sfm_export.h>
 
 #include <vital/algo/initialize_cameras_landmarks.h>
 
 namespace kwiver {
 namespace arrows {
-namespace core {
+namespace sfm {
 
 /// A class for initialization of cameras and landmarks
-class KWIVER_ALGO_CORE_EXPORT initialize_cameras_landmarks_keyframe
+class KWIVER_ALGO_SFM_EXPORT initialize_cameras_landmarks
 : public vital::algo::initialize_cameras_landmarks
 {
 public:
-  PLUGIN_INFO( "keyframe",
+  PLUGIN_INFO( "sfm",
                "Run SfM to estimate new cameras and landmarks "
-               "using feature tracks.  Gives keyframes priority." )
+               "using feature tracks." )
 
   /// Constructor
-  initialize_cameras_landmarks_keyframe();
+  initialize_cameras_landmarks();
 
   /// Destructor
-  virtual ~initialize_cameras_landmarks_keyframe();
+  virtual ~initialize_cameras_landmarks();
 
   /// Get this algorithm's \link vital::config_block configuration block \endlink
   virtual vital::config_block_sptr get_configuration() const;
@@ -92,7 +92,7 @@ private:
   const std::unique_ptr<priv> m_priv;
 };
 
-} // end namespace core
+} // end namespace sfm
 } // end namespace arrows
 } // end namespace kwiver
